requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
requests==2.32.3
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
rq==1.15.1
rq-scheduler==0.13.1
python-telegram-bot==20.7
streamlit==1.28.2
plotly==5.17.0
pandas==2.1.3
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
schedule==1.2.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
isort==5.12.0
discord.py==2.3.2
